System Information (with Expansion Port)
System information call updated for C256. Support included for the expansion cards.

+++ src/sys_general.h
@@ -52,25 +52,25 @@
  * Structure to describe the hardware
  */
 typedef struct s_sys_info {
-    unsigned short mcp_version;     /* Current version of the MCP kernel */
-    unsigned short mcp_rev;         /* Current revision, or sub-version of the MCP kernel */
-    unsigned short mcp_build;       /* Current vuild # of the MCP kernel */
-    unsigned short model;           /* Code to say what model of machine this is */
-    unsigned short sub_model;         /* 0x00 = PB, 0x01 = LB, 0x02 = CUBE */
+    uint16_t mcp_version;     		/* Current version of the MCP kernel */
+    uint16_t mcp_rev;         		/* Current revision, or sub-version of the MCP kernel */
+    uint16_t mcp_build;       		/* Current vuild # of the MCP kernel */
+    uint16_t model;           		/* Code to say what model of machine this is */
+    uint16_t sub_model;         	/* 0x00 = PB, 0x01 = LB, 0x02 = CUBE */
     const char * model_name;        /* Human readable name of the model of the computer */
-    unsigned short cpu;             /* Code to say which CPU is running */
+    uint16_t cpu;             		/* Code to say which CPU is running */
     const char * cpu_name;          /* Human readable name for the CPU */
-    unsigned int cpu_clock_khz;     /* Speed of the CPU clock in KHz */
+    uint32_t cpu_clock_khz;     	/* Speed of the CPU clock in KHz */
     unsigned long fpga_date;        /* YYYYMMDD */    
-    unsigned long fpga_model;       /* FPGA model number */
-    unsigned short fpga_version;    /* FPGA version */
-    unsigned short fpga_subver;     /* FPGA sub-version */
-    long system_ram_size;           /* The number of bytes of system RAM on the board */
+    uint16_t fpga_model;       		/* FPGA model number */
+    uint16_t fpga_version;    		/* FPGA version */
+    uint16_t fpga_subver;     		/* FPGA sub-version */
+    uint32_t system_ram_size;		/* The number of bytes of system RAM on the board */
     bool has_floppy;                /* TRUE if the board has a floppy drive installed */
     bool has_hard_drive;            /* TRUE if the board has a PATA device installed */
     bool has_expansion_card;        /* TRUE if an expansion card is installed on the device */
     bool has_ethernet;              /* TRUE if an ethernet port is present */
-    unsigned short screens;         /* How many screens are on this computer */
+    uint16_t screens;         		/* How many screens are on this computer */
 } t_sys_info, *p_sys_info;
 
 /*
